Rename internal header/source file according to conventions
