+++ cpp/src/arrow/compute/kernels/scalar_string_internal.h
